mexc-scalper/requirements.txt
PyQt6
pyqtgraph
qasync
websockets
numpy
pyyaml
fastapi
uvicorn
ccxt
python-dotenv
